format / lint / fix
issue: #326
